requirements.txt
pre-commit
fastapi
uvicorn 
pydantic
python-jose[cryptography] 
authlib
pydantic-settings
passlib
